feat: Autenticação rotas de CopyController, Atributo CopyCode único e validação das rotas de criar e editar Copy

+++ Controllers/CopiesController.cs
@@ -3,11 +3,12 @@ using ApiBiblioteca.Domain.DTOs;
 using ApiBiblioteca.Domain.Models.Interfaces;
 using ApiBiblioteca.Domain.Models;
 using AutoMapper;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Authorization;
 
 namespace ApiBiblioteca.Controllers
 {
+    [Authorize]
     [Route("api/v1/copies")]
     [ApiController]
     public class CopiesController : ControllerBase
@@ -39,28 +40,46 @@ namespace ApiBiblioteca.Controllers
             return copyDTO == null ? NotFound("Copy not found!") : Ok(copyDTO);
         }
 
-        [HttpPost]
-        public async Task<IActionResult> Add([FromBody] CopyViewModel viewModel)
+        [Authorize(Policy = "Admin")]
+        [HttpPost("book/{id}")]
+        public async Task<IActionResult> Add([FromBody] CopyViewModel viewModel, Guid id)
         {
+            var isRegisteredCopyCode = await _copyRepository.IsRegisteredCopyCode(viewModel.CopyCode);
+            if (isRegisteredCopyCode) return BadRequest("You already have a copy with this copycode registered.");
+
             var copy = _mapper.Map<Copy>(viewModel);
-            copy.BookId = Guid.Parse("95865670-26C3-4641-E29B-08DC3C4E5E48");
+            copy.BookId = id;
             await _copyRepository.Add(copy);
 
             return StatusCode(201, "Copy registered successfully!");
         }
 
+        [Authorize(Policy = "Admin")]
         [HttpPut("{id}")]
         public async Task<IActionResult> Update([FromBody] CopyViewModel viewModel, Guid id)
         {
+            var copyExists = await _copyRepository.GetById(id);
+            if (copyExists == null) return NotFound("Copy not found!");
+
+            var isRegisteredCopyCode = await _copyRepository.IsRegisteredCopyCode(viewModel.CopyCode);
+            if (isRegisteredCopyCode && copyExists.CopyCode != viewModel.CopyCode)
+            {
+                return BadRequest("You already have a copy with this copycode registered.");
+            }
+
             var copy = _mapper.Map<Copy>(viewModel);
             await _copyRepository.Update(id, copy);
 
             return Ok("Copy updated successfully!");
         }
 
+        [Authorize(Policy = "Admin")]
         [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(Guid id)
         {
+            var copyExists = await _copyRepository.GetById(id);
+            if (copyExists == null) return NotFound("Copy not found!");
+
             await _copyRepository.Delete(id);
 
             return Ok("Copy removed successfully!");
